move cli to `__main__.py`
